[refactor] handle fetch in CategoriesList Component
