+++ src/Components/Navbar.tsx
@@ -1,21 +1,8 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
-import { useHistory } from "react-router-dom";
 import { Link } from "react-router-dom";
+import CategoriesList from "./CategoriesList";
 
 export default function Navbar() {
-    let history = useHistory();
-    const [categories, setCategories] = useState([]);
-  
-    useEffect(() => {
-      axios
-        .get("https://api.chucknorris.io/jokes/categories")
-        .then((res) => setCategories(res.data));
-    }, []);
-
-    function handleSelect(event:any) {
-      history.push(`/jokes/category=${event.target.value}`);
-    }
+    
   
     return (
       <>
@@ -34,16 +21,7 @@ export default function Navbar() {
               <Link to="#about">About</Link>
             </li>
           </ul>
-          <div className="custom-select">
-            <span>Choose a category: </span>
-            <select onChange={handleSelect}>
-              {categories.map((category) => (
-                <option key={category} value={category}>
-                  {category}
-                </option>
-              ))}
-            </select>
-          </div>
+          <CategoriesList/>
         </div>
       </>
     );
